+++ src/main/java/com/sparta/bapzip/user/application/UserServiceV1.java
@@ -4,7 +4,6 @@ import com.sparta.bapzip.global.exception.ErrorCode;
 import com.sparta.bapzip.user.application.dto.request.UserDeleteRequestDto;
 import com.sparta.bapzip.user.application.dto.request.UserRoleChangeRequestDto;
 import com.sparta.bapzip.user.application.dto.request.UserUpdateRequestDto;
-import com.sparta.bapzip.user.application.dto.response.*;
 import com.sparta.bapzip.user.application.excpetion.DuplicateUserException;
 import com.sparta.bapzip.user.application.excpetion.PasswordNotMatchException;
 import com.sparta.bapzip.user.application.excpetion.UnauthorizedUserException;
@@ -13,6 +12,7 @@ import com.sparta.bapzip.user.domain.entity.UserEntity;
 import com.sparta.bapzip.user.domain.enums.UserRoleEnum;
 import com.sparta.bapzip.user.domain.repository.UserRepository;
 import com.sparta.bapzip.user.application.dto.request.SignupRequestDto;
+import com.sparta.bapzip.user.presentation.dto.response.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
