+++ src/main/java/com/sparta/bapzip/user/presentation/controller/UserControllerV1.java
@@ -6,12 +6,11 @@ import com.sparta.bapzip.user.application.dto.request.SignupRequestDto;
 import com.sparta.bapzip.user.application.dto.request.UserDeleteRequestDto;
 import com.sparta.bapzip.user.application.dto.request.UserRoleChangeRequestDto;
 import com.sparta.bapzip.user.application.dto.request.UserUpdateRequestDto;
-import com.sparta.bapzip.user.application.dto.response.*;
 import com.sparta.bapzip.user.domain.entity.UserDetailsImpl;
+import com.sparta.bapzip.user.presentation.dto.response.*;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
