+++ src/main/java/com/sparta/bapzip/user/presentation/dto/response/UserRoleChangeResponseDto.java
@@ -1,6 +1,7 @@
-package com.sparta.bapzip.user.application.dto.response;
+package com.sparta.bapzip.user.presentation.dto.response;
 
 import com.sparta.bapzip.user.domain.entity.UserEntity;
+import com.sparta.bapzip.user.domain.enums.UserRoleEnum;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -8,7 +9,7 @@ import java.time.LocalDateTime;
 
 @Getter
 @Builder
-public class UserUpdateResponseDto {
+public class UserRoleChangeResponseDto {
 
     private Long id;
 
@@ -16,16 +17,22 @@ public class UserUpdateResponseDto {
 
     private String name;
 
+    private UserRoleEnum role;
+
     private LocalDateTime updatedAt;
+
+    private Long updatedBy;
 
     private boolean isDeleted;
 
-    public static UserUpdateResponseDto of(UserEntity user) {
-        return UserUpdateResponseDto.builder()
+    public static UserRoleChangeResponseDto of(UserEntity user) {
+        return UserRoleChangeResponseDto.builder()
                 .id(user.getId())
                 .email(user.getEmail())
                 .name(user.getName())
+                .role(user.getRole())
                 .updatedAt(user.getUpdatedAt())
+                .updatedBy(user.getUpdatedBy())
                 .isDeleted(user.getIsDeleted())
                 .build();
     }
