+++ src/main/java/com/sparta/bapzip/user/presentation/dto/response/UserUpdateResponseDto.java
@@ -1,7 +1,6 @@
-package com.sparta.bapzip.user.application.dto.response;
+package com.sparta.bapzip.user.presentation.dto.response;
 
 import com.sparta.bapzip.user.domain.entity.UserEntity;
-import com.sparta.bapzip.user.domain.enums.UserRoleEnum;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -9,7 +8,7 @@ import java.time.LocalDateTime;
 
 @Getter
 @Builder
-public class UserRoleChangeResponseDto {
+public class UserUpdateResponseDto {
 
     private Long id;
 
@@ -17,22 +16,16 @@ public class UserRoleChangeResponseDto {
 
     private String name;
 
-    private UserRoleEnum role;
-
     private LocalDateTime updatedAt;
-
-    private Long updatedBy;
 
     private boolean isDeleted;
 
-    public static UserRoleChangeResponseDto of(UserEntity user) {
-        return UserRoleChangeResponseDto.builder()
+    public static UserUpdateResponseDto of(UserEntity user) {
+        return UserUpdateResponseDto.builder()
                 .id(user.getId())
                 .email(user.getEmail())
                 .name(user.getName())
-                .role(user.getRole())
                 .updatedAt(user.getUpdatedAt())
-                .updatedBy(user.getUpdatedBy())
                 .isDeleted(user.getIsDeleted())
                 .build();
     }
